+++ CHANGELOG.rst
@@ -4,6 +4,8 @@ Release 0.12.0 (unreleased)
 * Internal refactoring: introduce superproject & subproject (#896)
 * Switch from pykwalify to StrictYAML (#922)
 * Show line number when manifest validation fails (#36)
+* Add Fuzzing (#819)
+* Don't allow NULL or control characters in manifest (#114)
 
 Release 0.11.0 (released 2026-01-03)
 ====================================